ITI_TEST/requirements.txt
fastapi[all]
requests
requests-toolbelt
numpy
opencv-python
pillow
